// src/types/production/execution/shiftHandoverLog.ts
export interface ShiftHandoverLog {
  /**
   * 序号
   * 数字(0.) 只读
   */
  id: number;

  /**
   * 交接班日志编号
   * 字符(50) 只读
   * 规则：[业务标识]-[产线标识][工位标识]-[排班日期与班次]-[流水号] 的结构体系。
   * 业务标识：使用 JJB（交接班）作为固定前缀。
   * 产线标识：使用 ZY（再造烟叶）、GS（再造梗丝）、XJ（香精香料）。
   * 工位标识：提取核心词汇建立标准化的工位映射字典（例如：QCL, TQ, YCL 等）。
   * 排班日期与班次：采用 YYYYMMDD 结合单字符班次代码。使用 1（早班）、2（中班）、3（晚班）进行标识。
   * 流水号：使用2位数字（如 01）。该流水号在特定的“工位+班次”维度下递增。
   */
  logNo: string;

  /**
   * 交接班日志名称
   * 字符(50) 只读
   * 规则：[产线与记录表标准名]_[执行班组]_[日期班次]_牌号
   * 产线与记录表标准名如：再造烟叶前处理工序、再造梗丝原料预处理段、香精香料线等。
   * 执行班组：由“物理交接班日期（YYYYMMDD）”与“班次属性”拼接而成，如：早班、中班、晚班。
   */
  logName: string;

  /**
   * 生产任务编号
   * 字符(50) 只读
   * 规则：关联月度生产任务-月度生产安排-生产任务编号
   */
  taskNo: string;

  /**
   * 工序
   * 字符(50) 只读
   * 包含：再造烟叶前处理工序、再造梗丝原料预处理段、香精香料线等。
   */
  process: string;

  /**
   * 班组名称
   * 字符(20) 只读
   * 关联：班次主数据-班次信息-班组名称
   */
  teamName: string;

  /**
   * 班次名称
   * 字符(20) 只读
   * 关联：班次主数据-班次信息-班次名称
   */
  shiftName: string;

  /**
   * 提交人
   * 字符(20) 只读
   */
  submitter: string;

  /**
   * 提交时间
   * 时间（yyyy-mm-dd hh:mm:ss） 只读
   */
  submitTime: string;

  /**
   * 人工填报标识
   * 布尔 只读
   * 规则：如果页面通过异常转人工提交时，则需要体现此标识。
   */
  isManualFill: boolean;
}